transformer_transducer/requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.2.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.5.0
dill==0.3.8
docstring-to-markdown==0.15
exceptiongroup==1.2.2
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
grpcio==1.71.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.5.0
jedi==0.19.2
jedi-language-server==0.42.0
Jinja2==3.1.4
joblib==1.4.2
Levenshtein==0.27.1
lightning-utilities==0.14.2
lsprotocol==2023.0.1
Markdown==3.7
MarkupSafe==2.1.5
microvenv==2023.5.post1
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pygls==1.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchmetrics==1.7.0
torchvision==0.21.0+cu126
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.20.2
